severe modification of motion function
To allow proper functioning of touchdown function, motion function for valencia gantry have been modifed. Previous iteration had embedded a function to wait until the end of the motion. This was impeeding the reding othe current during the motion. The function has been removed and added as a separated tool. Need to be verified that all the other utility of the code, as focus-finding, fiducial-finding, etc. are working properly.